+++ src/languageSupport/languages/flux/lsp/utils.ts
@@ -1,0 +1,63 @@
+import {
+  ProtocolRequestType,
+  ProtocolNotificationType,
+} from 'vscode-languageserver-protocol'
+
+export enum Events {
+  LaunchWorker = 'Launch dedicated worker',
+  MonacoClientStarted = 'LanguageClient started',
+  MainThreadUp = 'MainThread is ready',
+  LspServerUp = 'LSP server is online',
+  WorkerThreadUp = 'WorkerThread is ready',
+}
+
+export const isJsonRpc = (msg: string): boolean => {
+  return JSON.stringify(msg).includes('jsonrpc')
+}
+
+export const makeRequestType = (method: string) => {
+  return new ProtocolRequestType<any, any, any, any, any>(method)
+}
+
+export const makeNotificationType = (method: string) => {
+  return new ProtocolNotificationType<any, any>(method)
+}
+
+const JSONRPC = '2.0',
+  FLUXLANGID = 'flux'
+
+const createNotification = (method: string, params: object = {}) => {
+  return {
+    jsonrpc: JSONRPC,
+    method,
+    params,
+  }
+}
+
+export const didOpen = (uri: string, text: string, version: number) => {
+  return createNotification(Methods.didOpen, {
+    textDocument: {
+      uri,
+      languageId: FLUXLANGID,
+      version,
+      text,
+    },
+  })
+}
+
+export enum Methods {
+  Initialize = 'initialize',
+  Initialized = 'initialized',
+  CompletionResolve = 'completionItem/resolve',
+  Completion = 'textDocument/completion',
+  Definition = 'textDocument/definition',
+  didChange = 'textDocument/didChange',
+  didOpen = 'textDocument/didOpen',
+  didSave = 'textDocument/didSave',
+  Highlight = 'textDocument/documentHighlight',
+  Symbol = 'textDocument/documentSymbol',
+  FoldingRange = 'textDocument/foldingRange',
+  Hover = 'textDocument/hover',
+  References = 'textDocument/references',
+  Rename = 'textDocument/rename',
+}
